Refactor migration code
