+++ homeassistant/components/unifiprotect/__init__.py
@@ -17,11 +17,10 @@ from homeassistant.const import (
     CONF_USERNAME,
     CONF_VERIFY_SSL,
     EVENT_HOMEASSISTANT_STOP,
-    Platform,
 )
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryAuthFailed, ConfigEntryNotReady
-from homeassistant.helpers import device_registry as dr, entity_registry as er
+from homeassistant.helpers import device_registry as dr
 from homeassistant.helpers.aiohttp_client import async_create_clientsession
 
 from .const import (
@@ -37,66 +36,13 @@ from .const import (
 )
 from .data import ProtectData, async_ufp_instance_for_config_entry_ids
 from .discovery import async_start_discovery
+from .migrate import async_migrate_data
 from .services import async_cleanup_services, async_setup_services
 from .utils import _async_unifi_mac_from_hass, async_get_devices
 
 _LOGGER = logging.getLogger(__name__)
 
 SCAN_INTERVAL = timedelta(seconds=DEFAULT_SCAN_INTERVAL)
-
-
-async def _async_migrate_data(
-    hass: HomeAssistant, entry: ConfigEntry, protect: ProtectApiClient
-) -> None:
-
-    registry = er.async_get(hass)
-    to_migrate = []
-    for entity in er.async_entries_for_config_entry(registry, entry.entry_id):
-        if entity.domain == Platform.BUTTON and "_" not in entity.unique_id:
-            _LOGGER.debug("Button %s needs migration", entity.entity_id)
-            to_migrate.append(entity)
-
-    if len(to_migrate) == 0:
-        _LOGGER.debug("No entities need migration")
-        return
-
-    _LOGGER.info("Migrating %s reboot button entities ", len(to_migrate))
-    bootstrap = await protect.get_bootstrap()
-    count = 0
-    for button in to_migrate:
-        device = None
-        for model in DEVICES_THAT_ADOPT:
-            attr = f"{model.value}s"
-            device = getattr(bootstrap, attr).get(button.unique_id)
-            if device is not None:
-                break
-
-        if device is None:
-            continue
-
-        new_unique_id = f"{device.id}_reboot"
-        _LOGGER.debug(
-            "Migrating entity %s (old unique_id: %s, new unique_id: %s)",
-            button.entity_id,
-            button.unique_id,
-            new_unique_id,
-        )
-        try:
-            registry.async_update_entity(button.entity_id, new_unique_id=new_unique_id)
-        except ValueError:
-            _LOGGER.warning(
-                "Could not migrate entity %s (old unique_id: %s, new unique_id: %s)",
-                button.entity_id,
-                button.unique_id,
-                new_unique_id,
-            )
-        else:
-            count += 1
-
-    if count < len(to_migrate):
-        _LOGGER.warning("Failed to migate %s reboot buttons", len(to_migrate) - count)
-    else:
-        _LOGGER.info("Migrated %s reboot button entities", count)
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
@@ -133,7 +79,7 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         )
         return False
 
-    await _async_migrate_data(hass, entry, protect)
+    await async_migrate_data(hass, entry, protect)
     if entry.unique_id is None:
         hass.config_entries.async_update_entry(entry, unique_id=nvr_info.mac)
